refactor: 💡 relocate unlimited usage math to cusEntsToUnlimitedUsage
